Twig and CrsfManager Modification
