+++ src/JHD/Framework/CsrfManager.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Created by PhpStorm.
+ * User: Xtrazyx
+ * Date: 01/09/2017
+ * Time: 15:26
+ */
+
+namespace JHD\Framework;
+
+use Symfony\Component\HttpFoundation\Session\Session;
+use Symfony\Component\Security\Csrf\CsrfTokenManager;
+use Symfony\Component\Security\Csrf\TokenGenerator\UriSafeTokenGenerator;
+use Symfony\Component\Security\Csrf\TokenStorage\SessionTokenStorage;
+
+abstract class CsrfManager
+{
+    static function createCsrfManager()
+    {
+        // CSRF Token Manager
+        $session = new Session();
+        $csrfGenerator = new UriSafeTokenGenerator();
+        $csrfStorage = new SessionTokenStorage($session);
+        $csrfManager = new CsrfTokenManager($csrfGenerator, $csrfStorage);
+
+        return $csrfManager;
+    }
+}
