+++ src/JHD/Framework/Twig.php
@@ -8,22 +8,48 @@
 
 namespace JHD\Framework;
 
+use Symfony\Bridge\Twig\Extension\FormExtension;
+use Symfony\Bridge\Twig\Extension\TranslationExtension;
+use Symfony\Bridge\Twig\Form\TwigRenderer;
+use Symfony\Bridge\Twig\Form\TwigRendererEngine;
+use Symfony\Component\Translation\Loader\XliffFileLoader;
+use Symfony\Component\Translation\Translator;
+
 trait Twig
 {
-    // Twig rendering engine
-    /**
-     * @param $viewPath string
-     * @param $vars array
-     */
-    public function render($viewPath, $vars)
+    public function getTwig()
     {
         $config = new Config();
-        $twigTemplatePath = ROOT_DIR . $config->getConfigsByKey('twig')['templatePath'];
+        $csrfManager = CsrfManager::createCsrfManager();
+
+        // TWIG
         $twigOptions = $config->getConfigsByKey('twig')['options'];
-        $loader = new \Twig_Loader_Filesystem($twigTemplatePath);
+        $templatePath = ROOT_DIR . $config->getConfigsByKey('twig')['templatePath'];
+        $bridgePath = ROOT_DIR . $config->getConfigsByKey('twig')['bridgePath'];
+        $defaultFormThemePath = $config->getConfigsByKey('twig')['defaultFormThemePath'];
+
+        $loader = new \Twig_Loader_Filesystem(array(
+            $templatePath,
+            $bridgePath,
+        ));
         $twig = new \Twig_Environment($loader, $twigOptions);
 
-        $template = $twig->loadTemplate($viewPath);
-        echo $template->render($vars);
+        $formEngine = new TwigRendererEngine(array($defaultFormThemePath), $twig);
+        $twig->addRuntimeLoader(new \Twig_FactoryRuntimeLoader(array(
+            TwigRenderer::class => function () use ($formEngine, $csrfManager) {
+                return new TwigRenderer($formEngine, $csrfManager);
+            },
+        )));
+
+        // Translation
+        $translator = new Translator('fr');
+        $translator->addLoader('xlf', new XliffFileLoader());
+        $translator->addResource('xlf', ROOT_DIR . $config->getConfigsByKey('symfony')['formTranslationPath'], 'fr', 'validators');
+        $translator->addResource('xlf', ROOT_DIR . $config->getConfigsByKey('symfony')['validatorTranslationPath'], 'fr', 'validators');
+
+        $twig->addExtension(new FormExtension());
+        $twig->addExtension(new TranslationExtension($translator));
+
+        return $twig;
     }
 }
